Continue working on date range validation rule
