+++ app/src/main/res/layout/activity_date_range.xml
@@ -61,8 +61,10 @@
                 android:layout_height="wrap_content"
                 app:datePattern="MM/dd/yyyy"
                 app:datePickerModes="input"
+                app:minDate="10/10/2020"
+                app:maxDate="10/10/2026"
                 app:fieldName="date"
-                app:formatterMode="flexible"
+                app:formatterMode="strict"
                 app:hint="Date range"
                 app:imeOptions="actionNext"
                 app:inputType="date"
@@ -75,6 +77,7 @@
             android:id="@+id/vgsTilExpirationDate"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
+            android:layout_marginHorizontal="@dimen/margin_padding_material_small"
             android:layout_marginTop="20dp"
             android:layout_weight="1"
             app:boxBackgroundModes="outline"
